% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_links_with_offset.R
\name{binomial_lik}
\alias{binomial_lik}
\title{Binomial Log-Likelihood}
\usage{
binomial_lik(x, mu, num_obs)
}
\arguments{
\item{x}{Vector of observed successes.}

\item{mu}{Vector of predicted probabilities.}

\item{num_obs}{Number of trials for each observation.}
}
\value{
Negative log-likelihood value.
}
\description{
Computes the negative log-likelihood for the Binomial distribution.
}
