% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_extension_adam.R
\name{compute_standard_errors}
\alias{compute_standard_errors}
\title{Compute Standard Errors via Fisher Information}
\usage{
compute_standard_errors(
  X,
  y,
  lambda,
  family,
  beta,
  good_beta,
  dispersion,
  weights,
  offset
)
}
\description{
Computes the variance-covariance matrix of estimates from the Fisher Information Matrix.
}
\keyword{internal}
