% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_links_with_offset.R
\name{gaussian_lik}
\alias{gaussian_lik}
\title{Gaussian Log-Likelihood}
\usage{
gaussian_lik(x, mu, sigma)
}
\arguments{
\item{x}{Vector of observed values.}

\item{mu}{Vector of predicted means.}

\item{sigma}{Standard deviation.}
}
\value{
Negative log-likelihood value.
}
\description{
Computes the negative log-likelihood for the Gaussian distribution.
}
\keyword{internal}
