% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_extension_adam.R
\name{gradient_descent_update}
\alias{gradient_descent_update}
\title{Compute Gradient Descent Updates for Spot-GLM}
\usage{
gradient_descent_update(
  family,
  X,
  y,
  beta,
  lambda,
  offset,
  disp = 1e+08,
  weights = rep(1, length(y))
)
}
\description{
Computes gradients of the negative log-likelihood for supported GLM families.
}
\keyword{internal}
