% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_extension_adam.R
\name{initialize_beta_full}
\alias{initialize_beta_full}
\title{Full Beta Initialization via Spot-GLM}
\usage{
initialize_beta_full(
  y,
  X,
  lambda,
  family,
  beta_0,
  offset = rep(0, length(y)),
  weights = rep(1, length(y)),
  fix_coef = matrix(FALSE, ncol(X), ncol(lambda)),
  intercept_ind = NULL
)
}
\description{
Uses spot-level model to initialize all coefficients with or without intercept-only pass.
}
\keyword{internal}
