% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_extension_adam.R
\name{initialize_beta_full_single_cell}
\alias{initialize_beta_full_single_cell}
\title{Full Beta Initialization (Single-Cell)}
\usage{
initialize_beta_full_single_cell(
  y,
  X,
  lambda,
  sc_family,
  beta,
  offset = rep(0, length(y)),
  weights = rep(1, length(y)),
  fix_coef = matrix(FALSE, ncol(X), ncol(lambda)),
  CT = NULL,
  min_freq = 50,
  intercept_ind = NULL
)
}
\description{
Initializes all coefficients using single-cell GLMs with optional intercept-only models.
}
\keyword{internal}
