% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_extension_adam.R
\name{initialize_beta_intercept}
\alias{initialize_beta_intercept}
\title{Initialize Intercept Coefficients via Spot-GLM}
\usage{
initialize_beta_intercept(
  y,
  X,
  lambda,
  family,
  beta_0,
  offset = rep(0, length(y)),
  weights = rep(1, length(y)),
  fix_coef = matrix(FALSE, ncol(X), ncol(lambda)),
  CT = NULL,
  min_freq = 50,
  intercept_ind = 1
)
}
\description{
Uses spot-level model to estimate intercept terms per cell type.
}
\keyword{internal}
