% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_extension_adam.R
\name{initialize_fix_coef}
\alias{initialize_fix_coef}
\title{Initialize Fixed Coefficient Matrix}
\usage{
initialize_fix_coef(
  X,
  lambda,
  fix_coef = matrix(FALSE, ncol(X), ncol(lambda)),
  weights = NULL,
  min_deconv = 0.1,
  min_freq = 100,
  intercept_col = NULL
)
}
\description{
Automatically sets which coefficients to fix based on low signal or coverage.
}
\keyword{internal}
