% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_links_with_offset.R
\name{nb_lik}
\alias{nb_lik}
\title{Negative Binomial Log-Likelihood}
\usage{
nb_lik(x, mu, disp)
}
\arguments{
\item{x}{Vector of observed counts.}

\item{mu}{Vector of expected means.}

\item{disp}{Dispersion parameter.}
}
\value{
Negative log-likelihood value.
}
\description{
Computes the negative log-likelihood for the Negative Binomial distribution.
}
\keyword{internal}
