% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_links_with_offset.R
\name{poisson_lik}
\alias{poisson_lik}
\title{Poisson Log-Likelihood}
\usage{
poisson_lik(x, mu)
}
\arguments{
\item{x}{Vector of observed counts.}

\item{mu}{Vector of predicted means (Poisson rates).}
}
\value{
Negative log-likelihood value.
}
\description{
Computes the negative log-likelihood for the Poisson distribution.
}
