% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_optimal_K.R
\name{read_spatial_long_read}
\alias{read_spatial_long_read}
\title{Read Spatial Long-Read RNA-Seq Data}
\usage{
read_spatial_long_read()
}
\value{
A list containing:
\describe{
\item{coords}{Matrix of spatial coordinates.}
\item{regions}{Spatial region annotations.}
\item{deconv}{Cell type deconvolution matrix.}
\item{library_size}{Vector of library sizes per spot.}
\item{total_gene_expression}{Matrix of total gene expression.}
\item{isoform_expression}{Matrix of isoform-level expression.}
}
}
\description{
Loads spatial long-read RNA-seq data from a GitHub repository. Includes coordinates,
region annotations, deconvolution, library sizes, and expression matrices for genes and isoforms.
}
\examples{
\dontrun{
long_read_data <- read_spatial_long_read()
str(long_read_data)
}

}
