% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_extension_adam.R
\name{run_model}
\alias{run_model}
\title{Run Spot-GLM Model with Initialization and Optimization}
\usage{
run_model(
  y,
  X,
  lambda,
  family = "spot gaussian",
  beta_0 = NULL,
  fix_coef = NULL,
  offset = rep(0, length(y)),
  initialization = "full",
  min_deconv = 0.1,
  min_freq = 50,
  CT = NULL,
  weights = rep(1, length(y)),
  ct_cov_weights = rep(1, ncol(lambda)),
  n_epochs = 100,
  batch_size = 500,
  learning_rate = 1,
  max_diff = 1 - 1e-06,
  improvement_threshold = 1e-06,
  max_conv = 10,
  intercept_col = NULL,
  min_reads_per_1000 = 1
)
}
\arguments{
\item{y}{Response vector.}

\item{X}{Covariate matrix.}

\item{lambda}{Deconvolution matrix.}

\item{family}{GLM family.}

\item{beta_0}{Optional initial coefficients.}

\item{fix_coef}{Logical matrix of fixed coefficients.}

\item{offset}{Optional offset vector.}

\item{initialization}{Method for coefficient initialization: \code{"intercept"} or \code{"full"}.}

\item{min_deconv}{Minimum required cell type proportion.}

\item{min_freq}{Minimum frequency of covariate values.}

\item{CT}{Optional cell type labels.}

\item{weights}{Observation weights.}

\item{ct_cov_weights}{Cell type weights.}

\item{learning_rate}{Initial learning rate.}

\item{max_diff}{Likelihood improvement threshold.}

\item{intercept_col}{Optional column index for intercept.}

\item{max_gd_steps}{Max gradient descent steps.}
}
\value{
A list with final coefficients, standard errors, likelihood, convergence status, and diagnostics.
}
\description{
Full pipeline for fitting a Spot-GLM to a single response (e.g. gene), including initialization, coefficient filtering,
and optimization via gradient descent.
}
