% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_extension_adam.R
\name{run_model_parallel_mac}
\alias{run_model_parallel_mac}
\title{Parallelized Spot-GLM Model Fitting (macOS / Linux)}
\usage{
run_model_parallel_mac(
  Y,
  X,
  lambda,
  family = "spot gaussian",
  beta_0 = NULL,
  fix_coef = NULL,
  initialization = "full",
  G = 0.1,
  num_cores = 1,
  offset = NULL,
  min_deconv = 0.1,
  min_freq = 50,
  CT = NULL,
  weights = NULL,
  ct_cov_weights = NULL,
  n_epochs = 100,
  batch_size = 500,
  learning_rate = 1,
  max_diff = 1 - 1e-06,
  improvement_threshold = 1e-06,
  max_conv = 10,
  intercept_col = NULL
)
}
\description{
Fits a Spot-GLM model for multiple responses (e.g., genes) in parallel using memory-safe chunking
and \code{pbmclapply}, which relies on \code{mclapply}. Only available on Unix-based systems.
}
\details{
This version uses \code{pbmcapply::pbmclapply} for parallelism. On Windows systems,
please use \code{\link{run_spot_glm_windows}}.
}
