% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_extension_adam.R
\name{run_single_cell}
\alias{run_single_cell}
\title{Run Single-Cell GLM per Cell Type}
\usage{
run_single_cell(
  y,
  X,
  lambda,
  sc_family = "gaussian",
  offset = rep(0, length(y)),
  weights = rep(1, length(y)),
  fix_coef = matrix(FALSE, ncol(X), ncol(lambda)),
  CT = NULL,
  min_freq = 50
)
}
\description{
Fits individual GLMs for each cell type using pseudo-bulked or single-cell data.
}
\keyword{internal}
