% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_links_with_offset.R
\docType{data}
\name{spot_gaussian}
\alias{spot_gaussian}
\title{Custom GLM Object: Spot Gaussian}
\format{
A list of class-specific functions:
\describe{
\item{\code{link}}{Link and inverse link functions with derivatives.}
\item{\code{marginal_mu}}{Marginal mean function.}
\item{\code{original_class}}{Base GLM family used (Gaussian).}
\item{\code{predict}}{Predict means from \code{X}, \code{beta}, and \code{lambda}.}
\item{\code{likelihood}}{Negative log-likelihood function.}
}
}
\usage{
spot_gaussian
}
\description{
Defines a Gaussian-based GLM family object for spatial transcriptomics data,
incorporating link functions and gradient-based optimization for continuous-valued
expression deconvolution.
}
\keyword{datasets}
